modernize-requirements/evals/fixtures/eval-3-uv-old-deps/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
fonttools==4.57.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.5
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
Pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
threadpoolctl==3.6.0
tqdm==4.67.1
tzdata==2025.2
urllib3==2.4.0
